SwiGLU fixes

+++ stream/inputs/aie/workload/make_onnx_swiglu.py
@@ -39,29 +39,29 @@ def make_swiglu_workload(seq_len, embedding_dim, hidden_dim, in_dtype, out_dtype
     # Initializers (weights)
     w_left = helper.make_tensor(
         "weights_1",
-        TensorProto.FLOAT,
+        TensorProto.BFLOAT16,
         [embedding_dim, hidden_dim],
         np.zeros((embedding_dim, hidden_dim)),
     )
     w_left.ClearField("float_data")
     w_right = helper.make_tensor(
         "weights_2",
-        TensorProto.FLOAT,
+        TensorProto.BFLOAT16,
         [embedding_dim, hidden_dim],
         np.zeros((embedding_dim, hidden_dim)),
     )
     w_right.ClearField("float_data")
     w_down = helper.make_tensor(
         "weights_3",
-        TensorProto.FLOAT,
+        TensorProto.BFLOAT16,
         [hidden_dim, embedding_dim],
         np.zeros((hidden_dim, embedding_dim)),
     )
     w_down.ClearField("float_data")
 
     # I/O
-    inp = helper.make_tensor_value_info("input", TensorProto.FLOAT, [seq_len, embedding_dim])
-    out = helper.make_tensor_value_info("output", TensorProto.FLOAT, [seq_len, embedding_dim])
+    inp = helper.make_tensor_value_info("input", TensorProto.BFLOAT16, [seq_len, embedding_dim])
+    out = helper.make_tensor_value_info("output", TensorProto.BFLOAT16, [seq_len, embedding_dim])
 
     # Branch GEMMs
     gemm_left = helper.make_node(
